# ast & topdown: Remove defers.