--- a/topdown/eval.go
+++ b/topdown/eval.go
@@ -21,8 +21,9 @@ type queryIDFactory struct {
 
 // Note: The first call to Next() returns 0.
 func (f *queryIDFactory) Next() uint64 {
-	defer func() { f.curr++ }()
-	return f.curr
+	curr := f.curr
+	f.curr++
+	return curr
 }
 
 type eval struct {
@@ -930,58 +931,77 @@ func getSavePairs(x *ast.Term, b *bindings, result []savePair) []savePair {
 
 func (e *eval) saveExpr(expr *ast.Expr, b *bindings, iter unifyIterator) error {
 	e.saveStack.Push(expr, b, b)
-	defer e.saveStack.Pop()
 	e.traceSave(expr)
-	return iter()
+	err := iter()
+	e.saveStack.Pop()
+	return err
 }
 
 func (e *eval) savePluggedExprs(exprs []*ast.Expr, iter unifyIterator) error {
 	for _, expr := range exprs {
 		e.saveStack.Push(expr, nil, nil)
-		defer e.saveStack.Pop()
 		e.traceSave(expr)
 	}
-	return iter()
+	err := iter()
+	for i := 0; i < len(exprs); i++ {
+		e.saveStack.Pop()
+	}
+	return err
 }
 
 func (e *eval) saveUnify(a, b *ast.Term, b1, b2 *bindings, iter unifyIterator) error {
 	e.instr.startTimer(partialOpSaveUnify)
 	expr := ast.Equality.Expr(a, b)
+	pops := 0
 	if pairs := getSavePairs(a, b1, nil); len(pairs) > 0 {
+		pops += len(pairs)
 		for _, p := range pairs {
 			e.saveSet.Push([]*ast.Term{p.term}, p.b)
-			defer e.saveSet.Pop()
 		}
+
 	}
 	if pairs := getSavePairs(b, b2, nil); len(pairs) > 0 {
+		pops += len(pairs)
 		for _, p := range pairs {
 			e.saveSet.Push([]*ast.Term{p.term}, p.b)
-			defer e.saveSet.Pop()
 		}
 	}
 	e.saveStack.Push(expr, b1, b2)
-	defer e.saveStack.Pop()
 	e.traceSave(expr)
 	e.instr.stopTimer(partialOpSaveUnify)
-	return iter()
+	err := iter()
+
+	e.saveStack.Pop()
+	for i := 0; i < pops; i++ {
+		e.saveSet.Pop()
+	}
+
+	return err
 }
 
 func (e *eval) saveCall(declArgsLen int, terms []*ast.Term, iter unifyIterator) error {
 	expr := ast.NewExpr(terms)
 	// If call-site includes output value then partial eval must add vars in output
 	// position to the save set.
+
+	pops := 0
 	if declArgsLen == len(terms)-2 {
 		if pairs := getSavePairs(terms[len(terms)-1], e.bindings, nil); len(pairs) > 0 {
+			pops += len(pairs)
 			for _, p := range pairs {
 				e.saveSet.Push([]*ast.Term{p.term}, p.b)
-				defer e.saveSet.Pop()
 			}
 		}
 	}
 	e.saveStack.Push(expr, e.bindings, nil)
-	defer e.saveStack.Pop()
 	e.traceSave(expr)
-	return iter()
+	err := iter()
+
+	e.saveStack.Pop()
+	for i := 0; i < pops; i++ {
+		e.saveSet.Pop()
+	}
+	return err
 }
 
 func (e *eval) getRules(ref ast.Ref) (*ast.IndexResult, error) {
@@ -1145,21 +1165,25 @@ func (e evalBuiltin) eval(iter unifyIterator) error {
 	numDeclArgs := len(e.bi.Decl.Args())
 
 	e.e.instr.startTimer(evalOpBuiltinCall)
-	defer e.e.instr.stopTimer(evalOpBuiltinCall)
 
-	return e.f(e.bctx, operands, func(output *ast.Term) error {
+	err := e.f(e.bctx, operands, func(output *ast.Term) error {
 
 		e.e.instr.stopTimer(evalOpBuiltinCall)
-		defer e.e.instr.startTimer(evalOpBuiltinCall)
 
+		var err error
 		if len(operands) == numDeclArgs {
 			if output.Value.Compare(ast.Boolean(false)) != 0 {
-				return iter()
+				err = iter()
 			}
-			return nil
+		} else {
+			err = e.e.unify(e.terms[len(e.terms)-1], output, iter)
 		}
-		return e.e.unify(e.terms[len(e.terms)-1], output, iter)
+		e.e.instr.startTimer(evalOpBuiltinCall)
+		return err
 	})
+
+	e.e.instr.stopTimer(evalOpBuiltinCall)
+	return err
 }
 
 type evalFunc struct {
@@ -1691,13 +1715,11 @@ func (e evalVirtualPartial) partialEvalSupportRule(iter unifyIterator, rule *ast
 	child.traceEnter(rule)
 
 	e.e.saveStack.PushQuery(nil)
-	defer e.e.saveStack.PopQuery()
 
-	return child.eval(func(child *eval) error {
+	err := child.eval(func(child *eval) error {
 		child.traceExit(rule)
 
 		current := e.e.saveStack.PopQuery()
-		defer e.e.saveStack.PushQuery(current)
 		plugged := current.Plug(child.bindings)
 
 		var key, value *ast.Term
@@ -1720,8 +1742,11 @@ func (e evalVirtualPartial) partialEvalSupportRule(iter unifyIterator, rule *ast
 		})
 
 		child.traceRedo(rule)
+		e.e.saveStack.PushQuery(current)
 		return nil
 	})
+	e.e.saveStack.PopQuery()
+	return err
 }
 
 func (e evalVirtualPartial) evalTerm(iter unifyIterator, term *ast.Term, termbindings *bindings) error {
@@ -1935,13 +1960,11 @@ func (e evalVirtualComplete) partialEvalSupportRule(iter unifyIterator, rule *as
 	child.traceEnter(rule)
 
 	e.e.saveStack.PushQuery(nil)
-	defer e.e.saveStack.PopQuery()
 
-	return child.eval(func(child *eval) error {
+	err := child.eval(func(child *eval) error {
 		child.traceExit(rule)
 
 		current := e.e.saveStack.PopQuery()
-		defer e.e.saveStack.PushQuery(current)
 		plugged := current.Plug(child.bindings)
 
 		head := ast.NewHead(rule.Head.Name, nil, child.bindings.PlugNamespaced(rule.Head.Value, child.bindings))
@@ -1954,8 +1977,11 @@ func (e evalVirtualComplete) partialEvalSupportRule(iter unifyIterator, rule *as
 		})
 
 		child.traceRedo(rule)
+		e.e.saveStack.PushQuery(current)
 		return nil
 	})
+	e.e.saveStack.PopQuery()
+	return err
 }
 
 func (e evalVirtualComplete) evalTerm(iter unifyIterator, term *ast.Term, termbindings *bindings) error {