Notable improvements are:

benchmark                                 old ns/op      new ns/op      delta
BenchmarkAuthzForbidAuthn-8               32946          31448          -4.55%
BenchmarkAuthzForbidPath-8                100565         98802          -1.75%
BenchmarkAuthzForbidMethod-8              106440         106324         -0.11%
BenchmarkAuthzAllow10Paths-8              102167         100737         -1.40%
BenchmarkAuthzAllow100Paths-8             633225         613168         -3.17%
BenchmarkAuthzAllow1000Paths-8            5987919        5490712        -8.30%
BenchmarkRESTAuthzForbidAuthn-8           525042         516558         -1.62%
BenchmarkRESTAuthzForbidPath-8            617314         594506         -3.69%
BenchmarkRESTAuthzForbidMethod-8          637980         601462         -5.72%
BenchmarkRESTAuthzAllow10Paths-8          628099         596365         -5.05%
BenchmarkRESTAuthzAllow100Paths-8         1173669        1151987        -1.85%
BenchmarkRESTAuthzAllow1000Paths-8        6276713        6309364        +0.52%
BenchmarkPartialEval/1-8                  9266           9150           -1.25%
BenchmarkPartialEval/10-8                 9310           9191           -1.28%
BenchmarkPartialEval/100-8                10117          9636           -4.75%
BenchmarkPartialEval/1000-8               10381          10273          -1.04%
BenchmarkPartialEvalCompile/1-8           4037594        4089562        +1.29%
BenchmarkPartialEvalCompile/10-8          6696489        6658199        -0.57%
BenchmarkPartialEvalCompile/100-8         58103030       51637149       -11.13%
BenchmarkPartialEvalCompile/1000-8        3325901936     2651853641     -20.27%
BenchmarkInliningFullScan/1000-8          9009538        8538953        -5.22%
BenchmarkInliningFullScan/10000-8         98244775       97655227       -0.60%
BenchmarkInliningFullScan/300000-8        3122413496     3041652765     -2.59%

Signed-off-by: Teemu Koponen <koponen@styra.com>

<!--

Thanks for submitting a PR to OPA!

Before pressing 'Create pull request' please read the checklist below.

* All code changes should be accompanied with tests. If you are not
modifying any tests, just provide a short explanation of why updates
to tests are not necessary. In addition to helping catch bugs, tests
are extremely helpful in providing _context_ that explains how your
changes can be used.

* All changes to public APIs **must** be accompanied with
docs. Examples of public APIs include built-in functions,
config fields, and of course, exported Go types/functions/constants/etc.

* Commit messages should explain _why_ you made the changes, not what
you changed. Use active voice. Keep the subject line under 50
characters or so.

* All commits must be signed off by the author. If you are not
familiar with signing off, see CONTRIBUTING.md below.

For more information on contributing to OPA see:

* [CONTRIBUTING.md](https://github.com/open-policy-agent/opa/blob/master/CONTRIBUTING.md)
  for high-level contribution guidelines.

* [DEVELOPMENT.md](https://github.com/open-policy-agent/opa/blob/master/docs/devel/DEVELOPMENT.md)
  for development workflow and environment setup.

-->