@koponen-styra 2019-08-22
Note, this does change the semantics around the panic handling, which I do not know if you care: defers get invoked with panics, these non-defers not.